правки в задании чтение и показ записей из MySQL и функции в отдельный файл
